+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Default the value of `azure_deployment` on all Azure Drivers to the model the Driver is using.
 - Field `azure_ad_token` on all Azure Drivers is no longer serializable.
+- Default standard OpenAI and Azure OpenAI image query model to `gpt-4o`.
+- **BREAKING**: Updated OpenAI-based image query drivers to remove Vision from the name.
 
 ## [0.25.1] - 2024-05-15